lol bruh

+++ Minecraft2/Game/Chunk.h
@@ -8,7 +8,7 @@ public:
 	
 	void Populate();
 
-	unsigned int blocks[32][32][32];
+	unsigned short blocks[32][32][32];
 	glm::ivec3 position = {0,0,0};
 	bool needsUpdate= false;
 };
